+++ core/src/main/java/com/roscopeco/moxy/matchers/Matchers.java
@@ -25,6 +25,7 @@ package com.roscopeco.moxy.matchers;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.function.Predicate;
 
 import com.roscopeco.moxy.Moxy;
 import com.roscopeco.moxy.api.MoxyEngine;
@@ -3258,6 +3259,298 @@ public class Matchers {
     return null;
   }
 
+  /* ************ FUNCTION ************** */
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>byte</code> if the supplied
+   * function returns true for that <code>byte</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static byte matchesByte(final Predicate<Byte> predicate) {
+    return matchesByte(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>byte</code> if the supplied
+   * function returns true for that <code>byte</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static byte matchesByte(final MoxyEngine engine, final Predicate<Byte> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return 0;
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>char</code> if the supplied
+   * predicate returns true for that <code>char</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static char matchesChar(final Predicate<Character> predicate) {
+    return matchesChar(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>char</code> if the supplied
+   * predicate returns true for that <code>char</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static char matchesChar(final MoxyEngine engine, final Predicate<Character> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return 0;
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>short</code> if the supplied
+   * predicate returns true for that <code>short</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static short matchesShort(final Predicate<Short> predicate) {
+    return matchesShort(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>short</code> if the supplied
+   * predicate returns true for that <code>short</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static short matchesShort(final MoxyEngine engine, final Predicate<Short> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return 0;
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>int</code> if the supplied
+   * predicate returns true for that <code>int</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static int matchesInt(final Predicate<Integer> predicate) {
+    return matchesInt(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>int</code> if the supplied
+   * predicate returns true for that <code>int</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static int matchesInt(final MoxyEngine engine, final Predicate<Integer> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return 0;
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>long</code> if the supplied
+   * predicate returns true for that <code>long</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static long matchesLong(final Predicate<Long> predicate) {
+    return matchesLong(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>long</code> if the supplied
+   * predicate returns true for that <code>long</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static long matchesLong(final MoxyEngine engine, final Predicate<Long> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return 0;
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>float</code> if the supplied
+   * predicate returns true for that <code>float</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static float matchesFloat(final Predicate<Float> predicate) {
+    return matchesFloat(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>float</code> if the supplied
+   * predicate returns true for that <code>float</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static float matchesFloat(final MoxyEngine engine, final Predicate<Float> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return 0;
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>double</code> if the supplied
+   * predicate returns true for that <code>double</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static double matchesDouble(final Predicate<Double> predicate) {
+    return matchesDouble(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>double</code> if the supplied
+   * predicate returns true for that <code>double</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return zero (ignored).
+   * @since 1.0
+   *
+   */
+  public static double matchesDouble(final MoxyEngine engine, final Predicate<Double> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return 0;
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>boolean</code> if the supplied
+   * predicate returns true for that <code>boolean</code>. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return false (ignored).
+   * @since 1.0
+   *
+   */
+  public static boolean matchesBool(final Predicate<Boolean> predicate) {
+    return matchesBool(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match a primitive <code>boolean</code> if the supplied
+   * predicate returns true for that <code>boolean</code>. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return false (ignored).
+   * @since 1.0
+   *
+   */
+  public static boolean matchesBool(final MoxyEngine engine, final Predicate<Boolean> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return false;
+  }
+
+  /**
+   * <p>Create a matcher that will match an argument of type <code>T</code> if the supplied
+   * predicate returns true for that argument. The matcher is created in the default
+   * {@link MoxyEngine}.</p>
+   *
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return null (ignored).
+   * @since 1.0
+   *
+   * @param <T> The type this matcher will operate on
+   */
+  public static <T> T matches(final Predicate<T> predicate) {
+    return matches(Moxy.getMoxyEngine(), predicate);
+  }
+
+  /**
+   * <p>Create a matcher that will match an argument of type <code>T</code> if the supplied
+   * predicate returns true for that argument. The matcher is created in the specified
+   * {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param predicate The predicate that will determine the match.
+   *
+   * @return null (ignored).
+   * @since 1.0
+   *
+   * @param <T> The type this matcher will operate on
+   */
+  public static <T> T matches(final MoxyEngine engine, final Predicate<T> predicate) {
+    engine.registerMatcher(new PredicateMatcher<>(predicate));
+    return null;
+  }
+
   /* ************ CUSTOM ************** */
 
   /**
@@ -3769,6 +4062,39 @@ public class Matchers {
    */
   public static String startsWith(final MoxyEngine engine, final String string) {
     engine.registerMatcher(new StartsWithMatcher(string));
+    return null;
+  }
+
+  /* ************ STRING - CONTAINS ************** */
+
+  /**
+   * <p>Creates a matcher for <code>String</code> arguments that matches
+   * if the argument contains the specified <code>String</code>.
+   * The matcher is created in the default Moxy engine.</p>
+   *
+   * @param string The string that arguments must contain in order to match.
+   *
+   * @return null (ignored).
+   * @see #contains(MoxyEngine, String)
+   * @since 1.0
+   */
+  public static String contains(final String string) {
+    return contains(Moxy.getMoxyEngine(), string);
+  }
+
+  /**
+   * <p>Creates a matcher for <code>String</code> arguments that matches
+   * if the argument contains the specified <code>String</code>.
+   * The matcher is created in the specified {@link MoxyEngine}.</p>
+   *
+   * @param engine The {@link MoxyEngine} to which this matcher applies.
+   * @param string The string that arguments must contain in order to match.
+   *
+   * @return null (ignored).
+   * @since 1.0
+   */
+  public static String contains(final MoxyEngine engine, final String string) {
+    engine.registerMatcher(new ContainsMatcher(string));
     return null;
   }