+++ core/src/test/java/com/roscopeco/moxy/TestMoxyMultithreaded.java
@@ -21,8 +21,42 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.roscopeco.example.model;
 
-public final class FinalClass {
-  public String returnHello() { return "Hello"; }
+package com.roscopeco.moxy;
+
+import static org.assertj.core.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+import com.roscopeco.moxy.model.classmock.SimpleClass;
+
+public class TestMoxyMultithreaded {
+  @Test
+  public void testMoxyMockWhenThenReturnDifferentThreads() throws InterruptedException {
+    final SimpleClass mock = Moxy.mock(SimpleClass.class);
+    Moxy.when(() -> mock.returnHello()).thenReturn("MARKER");
+
+    final String[] r = new String[] { "NOTSET" };
+    final Thread t = new Thread(() -> r[0] = mock.returnHello());
+    t.start();
+    t.join();
+
+    assertThat(r[0]).isEqualTo("MARKER");
+  }
+
+  @Test
+  public void testMoxyMockWhenThenVerifyDifferentThreads() throws InterruptedException {
+    final SimpleClass mock = Moxy.mock(SimpleClass.class);
+
+    final Thread t1 = new Thread(() -> mock.returnHello());
+    final Thread t2 = new Thread(() -> mock.returnHello());
+
+    t1.start();
+    t2.start();
+
+    t1.join();
+    t2.join();
+
+    Moxy.assertMock(() -> mock.returnHello()).wasCalledTwice();
+  }
 }