+++ core/src/test/java/com/roscopeco/moxy/model/ClassWithOverridenEqualsHashcode.java
@@ -1,0 +1,13 @@
+package com.roscopeco.moxy.model;
+
+public class ClassWithOverridenEqualsHashcode {
+  @Override
+  public int hashCode() {
+    return super.hashCode();
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    return super.equals(obj);
+  }
+}